Adding .schemaDirectives object property
When working with apollo-server or things that embed it like
graphql-yoga, it is important support directive processing
using the engine provided by them. More information about
Apollo's implementation of directive processing can be found here

https://www.apollographql.com/docs/graphql-tools/schema-directives.html

With respect to schemata, the property .schemaDirectives exists and
values can be stored there. This allows passing of the schemata instance
in place of config normally passed into a apollo server makeExecutableSchema
call if that approach is desired.